tweak
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/trunk@37955 72102866-910b-0410-8b05-ffd578937521

+++ config/config.align.c
@@ -12,8 +12,8 @@ int main ()
 {
   char  *buf = malloc(30);
   void  *v;
-  int   *sp;
-  int   *sq;
+  short *sp;
+  short *sq;
   int   *ip;
   int   *iq;
   int   i;
